+++ src/Service/S3/src/Input/CreateMultipartUploadRequest.php
@@ -190,6 +190,15 @@ final class CreateMultipartUploadRequest extends Input
     private $SSEKMSEncryptionContext;
 
     /**
+     * Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS
+     * KMS (SSE-KMS). Setting this header to `true` causes Amazon S3 to use an S3 Bucket Key for object encryption with
+     * SSE-KMS.
+     *
+     * @var bool|null
+     */
+    private $BucketKeyEnabled;
+
+    /**
      * @var null|RequestPayer::*
      */
     private $RequestPayer;
@@ -254,6 +263,7 @@ final class CreateMultipartUploadRequest extends Input
      *   SSECustomerKeyMD5?: string,
      *   SSEKMSKeyId?: string,
      *   SSEKMSEncryptionContext?: string,
+     *   BucketKeyEnabled?: bool,
      *   RequestPayer?: RequestPayer::*,
      *   Tagging?: string,
      *   ObjectLockMode?: ObjectLockMode::*,
@@ -287,6 +297,7 @@ final class CreateMultipartUploadRequest extends Input
         $this->SSECustomerKeyMD5 = $input['SSECustomerKeyMD5'] ?? null;
         $this->SSEKMSKeyId = $input['SSEKMSKeyId'] ?? null;
         $this->SSEKMSEncryptionContext = $input['SSEKMSEncryptionContext'] ?? null;
+        $this->BucketKeyEnabled = $input['BucketKeyEnabled'] ?? null;
         $this->RequestPayer = $input['RequestPayer'] ?? null;
         $this->Tagging = $input['Tagging'] ?? null;
         $this->ObjectLockMode = $input['ObjectLockMode'] ?? null;
@@ -312,6 +323,11 @@ final class CreateMultipartUploadRequest extends Input
     public function getBucket(): ?string
     {
         return $this->Bucket;
+    }
+
+    public function getBucketKeyEnabled(): ?bool
+    {
+        return $this->BucketKeyEnabled;
     }
 
     public function getCacheControl(): ?string
@@ -535,6 +551,9 @@ final class CreateMultipartUploadRequest extends Input
         if (null !== $this->SSEKMSEncryptionContext) {
             $headers['x-amz-server-side-encryption-context'] = $this->SSEKMSEncryptionContext;
         }
+        if (null !== $this->BucketKeyEnabled) {
+            $headers['x-amz-server-side-encryption-bucket-key-enabled'] = $this->BucketKeyEnabled ? 'true' : 'false';
+        }
         if (null !== $this->RequestPayer) {
             if (!RequestPayer::exists($this->RequestPayer)) {
                 throw new InvalidArgument(sprintf('Invalid parameter "RequestPayer" for "%s". The value "%s" is not a valid "RequestPayer".', __CLASS__, $this->RequestPayer));
@@ -603,6 +622,13 @@ final class CreateMultipartUploadRequest extends Input
     public function setBucket(?string $value): self
     {
         $this->Bucket = $value;
+
+        return $this;
+    }
+
+    public function setBucketKeyEnabled(?bool $value): self
+    {
+        $this->BucketKeyEnabled = $value;
 
         return $this;
     }