+++ src/Service/S3/src/Result/CreateMultipartUploadOutput.php
@@ -69,6 +69,11 @@ class CreateMultipartUploadOutput extends Result
      */
     private $SSEKMSEncryptionContext;
 
+    /**
+     * Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).
+     */
+    private $BucketKeyEnabled;
+
     private $RequestCharged;
 
     public function getAbortDate(): ?\DateTimeImmutable
@@ -90,6 +95,13 @@ class CreateMultipartUploadOutput extends Result
         $this->initialize();
 
         return $this->Bucket;
+    }
+
+    public function getBucketKeyEnabled(): ?bool
+    {
+        $this->initialize();
+
+        return $this->BucketKeyEnabled;
     }
 
     public function getKey(): ?string
@@ -165,6 +177,7 @@ class CreateMultipartUploadOutput extends Result
         $this->SSECustomerKeyMD5 = $headers['x-amz-server-side-encryption-customer-key-md5'][0] ?? null;
         $this->SSEKMSKeyId = $headers['x-amz-server-side-encryption-aws-kms-key-id'][0] ?? null;
         $this->SSEKMSEncryptionContext = $headers['x-amz-server-side-encryption-context'][0] ?? null;
+        $this->BucketKeyEnabled = isset($headers['x-amz-server-side-encryption-bucket-key-enabled'][0]) ? filter_var($headers['x-amz-server-side-encryption-bucket-key-enabled'][0], \FILTER_VALIDATE_BOOLEAN) : null;
         $this->RequestCharged = $headers['x-amz-request-charged'][0] ?? null;
 
         $data = new \SimpleXMLElement($response->getContent());