+++ src/Service/S3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## NOT RELEASED
+
+### Added
+
+- AWS api-change: S3 adds support for multiple-destination replication, option to sync replica modifications; S3 Bucket Keys to reduce cost of S3 SSE with AWS KMS
+
 ## 1.6.0
 
 ### Added